requirements.txt
crewai>=1.9.0
crewai-tools>=1.9.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
